refactor: flatten ConversationSummary read model to conventional Phoenix (Slice 2b)
Merge the ReadModels.ConversationSummary DTO + its Ecto read schema into a single
KlassHero.Messaging.ConversationSummary, and move the summaries repository's reads +
the synchronous system-note write-through onto the KlassHero.Messaging context. The
ConversationSummaries projection GenServer stays under adapters/driven/projections/
(only its schema alias flips) — ProgramCatalog precedent.

- New lib/klass_hero/messaging/conversation_summary.ex (schema-as-struct, no changesets;
  the projection owns all writes). Field-parity held, so to_dto is dropped — reads return
  the struct directly.
- Move list/total-unread/has-system-note/context reads (bare) + write_system_note_token
  (context_span, with the seed-fallback preserved) onto KlassHero.Messaging
- De-DI list_conversations, get_total_unread_count, get_conversation_context,
  reply_privately_to_broadcast; drop 2 config keys + ConversationSummariesRepository alias
- Repoint ConversationSummaries + EnrolledChildren projections and the summary queries to
  the flattened module
- Delete read-model DTO, read schema, summaries repository + 2 ports
- Interaction sweep: drop the summaries-repo probe, add Messaging to the flattened-contexts note
- Merge 3 dead repo/query unit tests into conversation_summary_test.exs (public-API coverage)

Full suite green (4198 passed). Slice 2b of the Messaging flatten.
