+++ lib/klass_hero/messaging/application/queries/get_total_unread_count.ex
@@ -7,11 +7,6 @@ defmodule KlassHero.Messaging.Application.Queries.GetTotalUnreadCount do
   notification badges in the navigation.
   """
 
-  @conversation_summaries_reader Application.compile_env!(:klass_hero, [
-                                   :messaging,
-                                   :for_querying_conversation_summaries
-                                 ])
-
   @doc """
   Gets total unread count for a user.
 
@@ -25,6 +20,6 @@ defmodule KlassHero.Messaging.Application.Queries.GetTotalUnreadCount do
   """
   @spec execute(String.t()) :: non_neg_integer()
   def execute(user_id) do
-    @conversation_summaries_reader.get_total_unread_count(user_id)
+    KlassHero.Messaging.summaries_total_unread_count(user_id)
   end
 end
