+++ lib/klass_hero/messaging/application/queries/list_conversations.ex
@@ -9,11 +9,6 @@ defmodule KlassHero.Messaging.Application.Queries.ListConversations do
 
   require Logger
 
-  @conversation_summaries_reader Application.compile_env!(:klass_hero, [
-                                   :messaging,
-                                   :for_querying_conversation_summaries
-                                 ])
-
   @doc """
   Lists conversations for a user.
 
@@ -40,7 +35,7 @@ defmodule KlassHero.Messaging.Application.Queries.ListConversations do
   @spec execute(String.t(), keyword()) ::
           {:ok, [map()], boolean()}
   def execute(user_id, opts \\ []) do
-    {:ok, summaries, has_more} = @conversation_summaries_reader.list_for_user(user_id, opts)
+    {:ok, summaries, has_more} = KlassHero.Messaging.list_conversation_summaries_for_user(user_id, opts)
 
     enriched = Enum.map(summaries, &to_enriched_map/1)
