+++ lib/klass_hero/messaging/conversation_summary.ex
@@ -1,0 +1,62 @@
+defmodule KlassHero.Messaging.ConversationSummary do
+  @moduledoc """
+  Denormalized read model for inbox display: one row per (conversation, participant).
+
+  Conventional Phoenix — this Ecto schema IS the struct that flows through the
+  `KlassHero.Messaging` read API. It has no changesets: the `ConversationSummaries`
+  projection owns all writes (via `insert_all`/`update_all`/`change` fragments), and
+  reads return this struct directly. IDs are generated by the projection, not the DB.
+  """
+
+  use Ecto.Schema
+
+  @primary_key {:id, :binary_id, autogenerate: false}
+  @timestamps_opts [type: :utc_datetime]
+
+  schema "conversation_summaries" do
+    field :conversation_id, :binary_id
+    field :user_id, :binary_id
+    field :conversation_type, :string
+    field :provider_id, :binary_id
+    field :program_id, :binary_id
+    field :subject, :string
+    field :other_participant_name, :string
+    field :program_name, :string
+    field :participant_count, :integer, default: 0
+    field :latest_message_content, :string
+    field :latest_message_sender_id, :binary_id
+    field :latest_message_at, :utc_datetime
+    field :unread_count, :integer, default: 0
+    field :last_read_at, :utc_datetime
+    field :archived_at, :utc_datetime
+    field :has_attachments, :boolean, default: false
+    field :system_notes, :map, default: %{}
+    field :enrolled_child_names, {:array, :string}, default: []
+
+    timestamps()
+  end
+
+  @type t :: %__MODULE__{
+          id: String.t(),
+          conversation_id: String.t(),
+          user_id: String.t(),
+          conversation_type: String.t() | nil,
+          provider_id: String.t(),
+          program_id: String.t() | nil,
+          subject: String.t() | nil,
+          other_participant_name: String.t() | nil,
+          program_name: String.t() | nil,
+          participant_count: non_neg_integer(),
+          latest_message_content: String.t() | nil,
+          latest_message_sender_id: String.t() | nil,
+          latest_message_at: DateTime.t() | nil,
+          unread_count: non_neg_integer(),
+          last_read_at: DateTime.t() | nil,
+          archived_at: DateTime.t() | nil,
+          has_attachments: boolean(),
+          system_notes: map(),
+          enrolled_child_names: [String.t()],
+          inserted_at: DateTime.t() | nil,
+          updated_at: DateTime.t() | nil
+        }
+end
